feat(release): independent release pipelines for CLI and daemon binaries (#516, #517)
Wire the daemon into the top-level Taskfile build aggregator and split the
release pipeline into two independent semantic-release workflows
(.releaserc_cli.json + flow-deploy-release-cli.yaml and the daemon
counterparts). Each binary now has its own tag namespace
(solo-provisioner-v* and solo-provisioner-daemon-v*) and its own VERSION
file under pkg/version/{cli,daemon}/ -- CLI and daemon can be released
independently and at different cadences.

The pkg/version package is split into symmetric cli/ and daemon/
subpackages that each embed their own VERSION/COMMIT and register their
Info with the parent package at init() time, so shared code under
internal/... continues to read the running binary's version through
version.Get / Number / Commit without knowing which binary it belongs to.

Closes #516, #517.
Part of epic #498 -- Two-Binary Build Layout.
