+++ .releaserc_daemon.json
@@ -1,0 +1,45 @@
+{
+  "tagFormat": "solo-provisioner-daemon-v${version}",
+  "plugins": [
+    ["@semantic-release/commit-analyzer", { "preset": "conventionalcommits" }],
+    ["@semantic-release/release-notes-generator", { "preset": "conventionalcommits" }],
+    [
+      "@semantic-release/github",
+      {
+        "assets": [
+          { "path": "bin/solo-provisioner-daemon-linux-amd64",            "label": "solo-provisioner-daemon (linux/amd64)" },
+          { "path": "bin/solo-provisioner-daemon-linux-amd64.sha256",     "label": "solo-provisioner-daemon (linux/amd64) SHA256" },
+          { "path": "bin/solo-provisioner-daemon-linux-amd64.sha256.asc", "label": "solo-provisioner-daemon (linux/amd64) SHA256 signature" },
+          { "path": "bin/solo-provisioner-daemon-linux-amd64.asc",        "label": "solo-provisioner-daemon (linux/amd64) signature" },
+
+          { "path": "bin/solo-provisioner-daemon-linux-arm64",            "label": "solo-provisioner-daemon (linux/arm64)" },
+          { "path": "bin/solo-provisioner-daemon-linux-arm64.sha256",     "label": "solo-provisioner-daemon (linux/arm64) SHA256" },
+          { "path": "bin/solo-provisioner-daemon-linux-arm64.sha256.asc", "label": "solo-provisioner-daemon (linux/arm64) SHA256 signature" },
+          { "path": "bin/solo-provisioner-daemon-linux-arm64.asc",        "label": "solo-provisioner-daemon (linux/arm64) signature" }
+        ]
+      }
+    ]
+  ],
+  "releaseRules": [
+    { "breaking": true, "release": "minor" }
+  ],
+  "verifyRelease": [
+    ["@semantic-release/exec", { "cmd": "printf \"%s\" \"${nextRelease.version}\" > pkg/version/daemon/VERSION" }],
+    ["@semantic-release/exec", { "cmd": "task hash:daemon:all" }]
+  ],
+  "prepare": [
+    ["@semantic-release/exec", { "cmd": "task sign:daemon:all" }]
+  ],
+  "branches": [
+    { "name": "main" },
+    {
+      "name": "release/([0-9]+).([0-9]+)",
+      "channel": "${name.replace(/release\\//g, '').split('.')[0]}.${name.replace(/release\\//g, '').split('.')[1]}.x",
+      "range": "${name.replace(/release\\//g, '').split('.')[0]}.${name.replace(/release\\//g, '').split('.')[1]}.x"
+    },
+    { "name": "alpha/*", "prerelease": "alpha", "channel": "alpha" },
+    { "name": "ci/*",    "prerelease": "alpha", "channel": "alpha" },
+    { "name": "beta/*",  "prerelease": "beta",  "channel": "beta"  },
+    { "name": "rc/*",    "prerelease": "rc",    "channel": "rc"    }
+  ]
+}
