+++ cmd/cli/commands/root.go
@@ -24,7 +24,7 @@ import (
 	"github.com/hashgraph/solo-weaver/internal/workflows"
 	"github.com/hashgraph/solo-weaver/pkg/config"
 	"github.com/hashgraph/solo-weaver/pkg/models"
-	"github.com/hashgraph/solo-weaver/pkg/version"
+	version "github.com/hashgraph/solo-weaver/pkg/version/cli"
 	"github.com/joomcode/errorx"
 	"github.com/spf13/cobra"
 )
