+++ pkg/version/cli/generate_unix.go
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+
+package cli
+
+//go:generate chmod +x generate_version_unix.sh
+//go:generate ./generate_version_unix.sh
