+++ pkg/version/cli/generate_version_unix.sh
