+++ pkg/version/cli/generate_version_windows.ps1
