+++ pkg/version/cli/generate_windows.go
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: Apache-2.0
+
+package cli
+
+//go:generate ./generate_version_windows.ps1
