+++ pkg/version/cli/releases.go
@@ -1,0 +1,57 @@
+// SPDX-License-Identifier: Apache-2.0
+
+// Package cli is the version subpackage owned by the solo-provisioner CLI binary.
+//
+// It embeds the CLI's own VERSION and COMMIT files (generated by
+// pkg/version/cli/generate_version_unix.sh during `go generate`) and exposes
+// the canonical version accessors and Cobra "version" command for the CLI.
+package cli
+
+import (
+	_ "embed"
+	"runtime"
+
+	"github.com/spf13/cobra"
+
+	"github.com/hashgraph/solo-weaver/pkg/version"
+)
+
+//go:embed COMMIT
+var commit string
+
+//go:embed VERSION
+var number string
+
+func Commit() string {
+	return commit
+}
+
+func Number() string {
+	return number
+}
+
+// Get returns the CLI binary's version Info.
+func Get() version.Info {
+	return version.Info{
+		Number:    number,
+		Commit:    commit,
+		GoVersion: runtime.Version(),
+	}
+}
+
+// init registers the CLI binary's Info with the parent package so shared code
+// under internal/... (which imports pkg/version, not pkg/version/cli) can read
+// the running binary's version through version.Get / Number / Commit.
+func init() {
+	version.SetCurrent(Get())
+}
+
+// Cmd returns a "version" Cobra command bound to the CLI binary's Info.
+func Cmd() *cobra.Command {
+	return version.NewCmd(Get)
+}
+
+// Print writes the CLI binary's Info to cmd's stdout in the requested format.
+func Print(cmd *cobra.Command, format string) error {
+	return version.Print(cmd, format, Get())
+}
