+++ pkg/version/daemon/generate_unix.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
-package version
+package daemon
 
 //go:generate chmod +x generate_version_unix.sh
 //go:generate ./generate_version_unix.sh
