+++ pkg/version/daemon/generate_version_unix.sh
@@ -1,0 +1,11 @@
+#!/usr/bin/env bash
+# SPDX-License-Identifier: Apache-2.0
+
+if [[ ! -f VERSION ]]; then
+  echo -n "0.0.0" >VERSION
+  echo "Generated VERSION file with the default version...."
+fi
+
+COMMIT="$(git rev-parse HEAD)"
+echo -n "${COMMIT}" >COMMIT
+echo "Generated COMMIT file with the current commit hash...."
