+++ pkg/version/daemon/generate_windows.go
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
 
-package version
+package daemon
 
 //go:generate ./generate_version_windows.ps1
