+++ pkg/version/daemon/releases.go
@@ -1,0 +1,57 @@
+// SPDX-License-Identifier: Apache-2.0
+
+// Package daemon is the version subpackage owned by the solo-provisioner-daemon binary.
+//
+// It embeds the daemon's own VERSION and COMMIT files (generated by
+// pkg/version/daemon/generate_version_unix.sh during `go generate`) and exposes
+// the canonical version accessors and Cobra "version" command for the daemon.
+package daemon
+
+import (
+	_ "embed"
+	"runtime"
+
+	"github.com/spf13/cobra"
+
+	"github.com/hashgraph/solo-weaver/pkg/version"
+)
+
+//go:embed COMMIT
+var commit string
+
+//go:embed VERSION
+var number string
+
+func Commit() string {
+	return commit
+}
+
+func Number() string {
+	return number
+}
+
+// Get returns the daemon binary's version Info.
+func Get() version.Info {
+	return version.Info{
+		Number:    number,
+		Commit:    commit,
+		GoVersion: runtime.Version(),
+	}
+}
+
+// init registers the daemon binary's Info with the parent package so shared
+// code under internal/... (which imports pkg/version, not pkg/version/daemon)
+// can read the running binary's version through version.Get / Number / Commit.
+func init() {
+	version.SetCurrent(Get())
+}
+
+// Cmd returns a "version" Cobra command bound to the daemon binary's Info.
+func Cmd() *cobra.Command {
+	return version.NewCmd(Get)
+}
+
+// Print writes the daemon binary's Info to cmd's stdout in the requested format.
+func Print(cmd *cobra.Command, format string) error {
+	return version.Print(cmd, format, Get())
+}
