+++ pkg/version/info.go
@@ -1,11 +1,16 @@
 // SPDX-License-Identifier: Apache-2.0
 
+// Package version holds the shared types and helpers used by each binary's
+// version subpackage (pkg/version/cli and pkg/version/daemon). The two
+// subpackages each embed their own VERSION + COMMIT and register themselves
+// here at init time via SetCurrent so shared code (internal/...) can read the
+// running binary's version through Get / Number / Commit without knowing which
+// binary it belongs to.
 package version
 
 import (
 	"encoding/json"
 	"fmt"
-	"runtime"
 	"strings"
 
 	"github.com/joomcode/errorx"
@@ -52,18 +57,28 @@ func (v Info) Text() string {
 	return fmt.Sprintf("Version: %s\nCommit: %s\nGo Version: %s", v.Number, v.Commit, v.GoVersion)
 }
 
-var (
-	versionInfo Info
-)
+// current holds the running binary's Info, registered at init time by the
+// corresponding pkg/version/{cli,daemon} subpackage. Shared code under
+// internal/... reads it via Get / Number / Commit.
+var current Info
 
-func init() {
-	versionInfo = Info{
-		Number:    Number(),
-		Commit:    Commit(),
-		GoVersion: runtime.Version(),
-	}
+// SetCurrent registers the running binary's Info. Called from each binary's
+// pkg/version/{cli,daemon} subpackage init().
+func SetCurrent(info Info) {
+	current = info
 }
 
+// Get returns the running binary's Info.
 func Get() Info {
-	return versionInfo
+	return current
+}
+
+// Number returns the running binary's version number.
+func Number() string {
+	return current.Number
+}
+
+// Commit returns the running binary's commit hash.
+func Commit() string {
+	return current.Commit
 }
